+++ src/OSPSuite.Presentation/Views/Importer/IImporterView.cs
@@ -5,7 +5,7 @@ namespace OSPSuite.Presentation.Views.Importer
    public interface IImporterView : IView<IImporterPresenter>
    {
       void AddImporterView(IImporterDataView importerDataView);
-      void AddColumnMappingControl(IColumnMappingControl columnMappingControl);
+      void AddColumnMappingControl(IColumnMappingView columnMappingView);
       void AddConfirmationView(IImportConfirmationView confirmationView);
       void AddSourceFileControl(ISourceFileControl sourceFileControl);
       void EnableConfirmationView();
