+++ src/OSPSuite.UI/Views/Importer/MappingParameterEditorView.cs
@@ -1,5 +1,6 @@
 ﻿using OSPSuite.Presentation.Presenters.Importer;
 using OSPSuite.Presentation.Views;
+using OSPSuite.Presentation.Views.Importer;
 using OSPSuite.UI.Controls;
 using OSPSuite.UI.Extensions;
